Remove old classes
