--- VMagicMirror/Assets/Baku/VMagicMirror/Scripts/AvatarControl/WordToMotion/GamepadToWordToMotion.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 793cdd0b156d6764dbdef193411de76c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
