--- VMagicMirror/Assets/Baku/VMagicMirror/Scripts/AvatarControl/WordToMotion/MidiToWordToMotion.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fdd2c3b70407e974998d3dd5465feb39
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
