+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/AvatarControl/WordToMotion/V2/RequestSource/MidiNoteToMotionMap.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1a27d6ae3a3e79845940da8cc571003b
+guid: 2e6243683e04ffa44a96f3587332c44b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
